backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
django-filter
markdown
django-import-export
openpyxl
whitenoise
drf-spectacular
python-dateutil
